requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
aiosqlite==0.19.0
cryptography>=42.0.0
httpx[http2]==0.26.0
qrcode[pil]==7.4.2
pillow>=10.0.0
python-multipart==0.0.6
aioapns==3.2
